renaming routing test files
